+++ client/src/languageFeature/compile.ts
@@ -87,14 +87,14 @@ function getIgnoreRegular(languageId:string) {
 	return regular;
 }
 
-export async function compileBlock(block:ICommentBlock,languageId:string): Promise<ITranslatedText> {
+export async function compileBlock(block:ICommentBlock,languageId:string,targetLanguage?:string): Promise<ITranslatedText> {
 
 	let translatedText: string;
 	let humanizeText: string = '';
 	const { comment: originText } = block;
 	let { tokens } = block;
 
-	const targetLanguage = getConfig<string>('targetLanguage') || userLanguage;
+	targetLanguage = targetLanguage || getConfig<string>('targetLanguage') || userLanguage;
 	if (!tokens) {
 		// 选取翻译&单个单词翻译的时候。无tokens的简单结果
 		humanizeText = humanize(originText);